+++ CHANGES.rst
@@ -42,7 +42,15 @@ astropy.modeling
 ^^^^^^^^^^^^^^^^
 
 - Added NFW profile and tests to modeling package [#10505]
+
 - Added missing logic for evaluate to compound models [#10002]
+
+- Stop iteration in ``FittingWithOutlierRemoval`` before reaching ``niter`` if
+  the masked points are no longer changing. [#10642]
+
+- Keep a (shallow) copy of ``fit_info`` from the last iteration of the wrapped
+  fitter in ``FittingWithOutlierRemoval`` and also record the actual number of
+  iterations performed in it. [#10642]
 
 astropy.nddata
 ^^^^^^^^^^^^^^